assistant_orchestrator/requirements.txt
pyautogen==0.2.17
openai==1.13.3